Update ModelStub subclasses
